backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
PyPDF2
nltk
scikit-learn
numpy
pandas
cors